Proper remove script handling and some other things!  :D
• Remove scripts are extracted from *all* script files and merged when
Merge Scripts is enabled! (not just the install script being merged!)

• When decompiling PRAY files, the remove script is also copied to a
file just like "Script 1" and etc. for convenience to help with giant
remove scripts.  Note that it is NOT read when recompiling the agent and
only the normal "Remove script" tag in the praysource.txt is used like
it always was (since the .txt pray source format doesn't support @ .cos
files for "Remove script" (as far as I know), and I don't control that
format.  CAOS2PRAY has always supported external remove script files
with *#rscr though :3 )

• UTF-8 support all throughout!
• Universal Newlines done consistently now!
• Minor things